Fibonacci4

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,16 +8,16 @@
 
 int main(void)
 {
-	unsigned int fib1 = 1, fib2 = 2, next;
+	long double fib1 = 1, fib2 = 2, next;
 	int count;
 
-	printf("%u, %u", fib1, fib2);
+	printf("%.0LF, %0LF", fib1, fib2);
 
 	for (count = 3; count <= 98; count++)
 	{
 	next = fib1 + fib2;
 
-	printf(", %u", next);
+	printf(", %.0LF", next);
 
 	fib1 = fib2;
 	fib2 = next;
